Don’t score and sort all captures in RECAPTURES stage.
For these recaptures, we’re only considering those captures that recapture the recapture square (small portion of all the captures). Therefore, scoring all of the captures and pick_besting out of the whole group is not necessary.

All credit to original simplification in SF Dev goes to protonspring:

protonspring/Stockfish@5fb5beb